EDU-456 feat: 추천 컨텐츠 이벤트 구현
