+++ docker-compose.yml
@@ -57,7 +57,7 @@ services:
       KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181 # 브로커가 메타데이터를 저장할 ZK
       KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1 # 컨슈머 그룹 오프셋 토픽의 복제계수, 싱글 노드라 1 운영은 3 권장
       # ------ 리스너 설정 ------
-      KAFKA_LISTENERS: INTERNAL://0.0.0.0:9092,EXTERNAL://0.0.0.0:19092
+      KAFKA_LISTENERS: EXTERNAL://0.0.0.0:9092,INTERNAL://0.0.0.0:19092
       KAFKA_ADVERTISED_LISTENERS: EXTERNAL://localhost:9092,INTERNAL://kafka:19092
       KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: INTERNAL:PLAINTEXT,EXTERNAL:PLAINTEXT
       KAFKA_INTER_BROKER_LISTENER_NAME: INTERNAL
