+++ src/main/java/com/education/takeit/TakeitApplication.java
@@ -8,7 +8,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
 @EnableCaching
@@ -18,7 +17,6 @@ import org.springframework.scheduling.annotation.EnableAsync;
   KakaoProperties.class,
   NaverProperties.class
 })
-@EnableAsync
 public class TakeitApplication {
   public static void main(String[] args) {
     SpringApplication.run(TakeitApplication.class, args);
