--- src/main/java/com/education/takeit/kafka/dto/FeedbackFailDto.java
@@ -1,4 +1,0 @@
-package com.education.takeit.kafka.dto;
-
-public record FeedbackFailDto(
-    Long userId, Long subjectId, String type, int nth, String errorCode, String errorMessage) {}
