--- src/main/java/com/education/takeit/kafka/dto/FeedbackRequestDto.java
@@ -1,3 +1,0 @@
-package com.education.takeit.kafka.dto;
-
-public record FeedbackRequestDto(Long userId, Long subjectId, String type, int nth) {}
