--- src/main/java/com/education/takeit/kafka/dto/FeedbackResultDto.java
@@ -1,6 +1,0 @@
-package com.education.takeit.kafka.dto;
-
-import com.education.takeit.feedback.dto.FeedbackResponseDto;
-
-public record FeedbackResultDto(
-    Long userId, Long subjectId, String type, int nth, FeedbackResponseDto feedback) {}
