+++ src/main/java/com/education/takeit/kafka/feedback/dto/FeedbackFailDto.java
@@ -1,0 +1,4 @@
+package com.education.takeit.kafka.feedback.dto;
+
+public record FeedbackFailDto(
+    Long userId, Long subjectId, String type, int nth, String errorCode, String errorMessage) {}
