+++ src/main/java/com/education/takeit/kafka/feedback/dto/FeedbackRequestDto.java
@@ -1,0 +1,3 @@
+package com.education.takeit.kafka.feedback.dto;
+
+public record FeedbackRequestDto(Long userId, Long subjectId, String type, int nth) {}
