+++ src/main/java/com/education/takeit/kafka/recommand/dto/RecomFailDto.java
@@ -1,0 +1,9 @@
+package com.education.takeit.kafka.recommand.dto;
+
+public record RecomFailDto(
+        Long userId,
+        Long subjectId,
+        String errorCode,
+        String errorMessage
+) {
+}
