+++ src/main/java/com/education/takeit/kafka/recommand/dto/RecomRequestDto.java
@@ -1,0 +1,7 @@
+package com.education.takeit.kafka.recommand.dto;
+
+public record RecomRequestDto(
+        Long userId,
+        Long subjectId
+) {
+}
