+++ src/main/java/com/education/takeit/kafka/recommand/dto/RecomResultDto.java
@@ -1,0 +1,12 @@
+package com.education.takeit.kafka.recommand.dto;
+
+import com.education.takeit.recommend.dto.UserContentResDto;
+
+import java.util.List;
+
+public record RecomResultDto(
+        Long userId,
+        Long subjectId,
+        List<UserContentResDto> recommendation
+) {
+}
