+++ src/main/java/com/education/takeit/kafka/recommand/repository/RecomFailRepository.java
@@ -1,0 +1,7 @@
+package com.education.takeit.kafka.recommand.repository;
+
+import com.education.takeit.kafka.recommand.entity.RecomFailLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecomFailRepository extends JpaRepository<RecomFailLog, Long> {
+}
