--- src/main/java/com/education/takeit/kafka/repository/FeedbackFailRepository.java
@@ -1,6 +1,0 @@
-package com.education.takeit.kafka.repository;
-
-import com.education.takeit.kafka.entity.FeedbackFailLog;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FeedbackFailRepository extends JpaRepository<FeedbackFailLog, Long> {}
