+++ src/main/java/com/education/takeit/recommend/service/RecommendService.java
@@ -2,22 +2,33 @@ package com.education.takeit.recommend.service;
 
 import com.education.takeit.exam.service.ExamService;
 import com.education.takeit.global.client.AIClient;
+import com.education.takeit.global.dto.StatusCode;
+import com.education.takeit.global.exception.CustomException;
+import com.education.takeit.kafka.recommand.dto.RecomResultDto;
 import com.education.takeit.recommend.dto.UserContentResDto;
 import com.education.takeit.recommend.entity.TotalContent;
 import com.education.takeit.recommend.entity.UserContent;
+import com.education.takeit.recommend.repository.TotalContentRepository;
 import com.education.takeit.recommend.repository.UserContentRepository;
-import java.util.List;
-import java.util.stream.Collectors;
+import com.education.takeit.roadmap.entity.Subject;
+import com.education.takeit.roadmap.repository.SubjectRepository;
+import com.education.takeit.user.entity.User;
+import com.education.takeit.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class RecommendService {
   private final UserContentRepository userContentRepository;
+  private final UserRepository userRepository;
+  private final SubjectRepository subjectRepository;
+  private final TotalContentRepository totalContentRepository;
   private final ExamService examService;
   private final AIClient aiClient;
 
@@ -67,21 +78,33 @@ public class RecommendService {
         .collect(Collectors.toList());
   }
 
-  // 추천 컨텐츠 요청
-  @Async // FastAPI로 요청을 보내고 응답 받는 것은 비동기로 처리. 응답을 생성하는데 시간이 오래 걸리기 때문.
-  public void fetchAndSaveRecommendation(Long userId, Long subjectId) {
-    try {
-      List<UserContentResDto> recommendationList = aiClient.getRecommendation(userId, subjectId);
-      examService.saveUserContent(userId, recommendationList);
-      log.info("추천 컨텐츠 저장 완료!!!!!!!!!!!!!!! userId  : {}, subjectId  : {}", userId, subjectId);
-    } catch (Exception e) {
-      // 도메인 로직 중 발생한 예외
-      log.warn(
-          "추천 컨텐츠 저장 실패!!!!!!!!!!!!!!!!!! - userId: {}, subjectId: {}, reason: {}",
-          userId,
-          subjectId,
-          e.getMessage(),
-          e);
-    }
+  public void saveUserContents(RecomResultDto dto) {
+      User user =
+              userRepository
+                      .findByUserId(dto.userId())
+                      .orElseThrow(() -> new CustomException(StatusCode.NOT_EXIST_USER));
+      Subject subject =
+              subjectRepository
+                      .findBySubId(dto.subjectId())
+                      .orElseThrow(() -> new CustomException(StatusCode.NOT_EXIST_SUBJECT));
+
+      List<UserContent> recommendations =
+              dto.recommendation().stream()
+                      .map(
+                              userContentResDto -> {
+                                  TotalContent totalContent =
+                                          totalContentRepository
+                                                  .findById(userContentResDto.contentId())
+                                                  .orElseThrow(() -> new CustomException(StatusCode.CONTENTS_NOT_FOUND));
+                                  return UserContent.builder()
+                                          .totalContent(totalContent)
+                                          .subject(subject)
+                                          .user(user)
+                                          .isAiRecommended(userContentResDto.isAiRecommendation())
+                                          .aiRecommendReason(userContentResDto.comment())
+                                          .build();
+                              }
+                      ).toList();
+      userContentRepository.saveAll(recommendations);
   }
 }
